recipes/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==4.1
djangorestframework==3.14
djongo==1.3.7
dnspython==2.7.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pymongo==3.11.4
pytest==8.3.5
pytest-django==4.10.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.2.4
uritemplate==4.1.1
